add blueprint & complete ajax
