+++ templates/index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet"  type= "text/css" href="static/css/index.css?v=0.0.20">
+    <link rel="shortcut icon" href="#"></link>
     <title>首頁</title>
 </head>
 <body onload="ajax('/api/attractions/')">
@@ -52,6 +53,6 @@
 
 
 
-<script type="text/javascript" src="../static/js/index.js?v=37"></script>
+<script type="text/javascript" src="../static/js/index.js?v=41"></script>
 </body>
 </html>
