+++ templates/thankyou.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="shortcut icon" href="#"></link>
+    <title>Document</title>
+</head>
+<body>
+test
+<script type="text/javascript" src="../static/js/1.js?v=37"></script>   
+</body>
+</html>
